wolf

+++ wolfenstein3d/srcs/utils.c
@@ -6,28 +6,28 @@
 /*   By: midrissi <midrissi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/31 14:18:44 by midrissi          #+#    #+#             */
-/*   Updated: 2019/03/31 14:35:55 by midrissi         ###   ########.fr       */
+/*   Updated: 2019/03/31 15:37:44 by midrissi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "wolf3d.h"
 
-static float	toradian(float angle)
+static double	toradian(double angle)
 {
 	return (angle * M_PI / 180.0);
 }
 
-float			tcos(float angle)
+double			tcos(double angle)
 {
 	return (cos(toradian(angle)));
 }
 
-float			tsin(float angle)
+double			tsin(double angle)
 {
 	return (sin(toradian(angle)));
 }
 
-float			ttan(float angle)
+double			ttan(double angle)
 {
 	return (tan(toradian(angle)));
 }
